# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SugarRush API",
    "description": "The SugarRush API enables candy distributors to manage inventory, store accounts, and deliveries efficiently. It provides endpoints for accessing candy data, managing store profiles and orders, and tracking deliveries. The API also includes user authentication via Google OAuth.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/users/logout": {
      "post": {
        "tags": ["users"],
        "summary": "Log out the user",
        "description": "Logs the user out of Google, invalidates the current access token, and blacklists the JWT via Redis.\n**Important:** After logout, click \"Authorize\" → \"Logout\" to clear Swagger's stored token.",
        "operationId": "userLogout",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Successfully logged out" },
          "400": { "description": "No token provided or invalid token" },
          "401": { "description": "Unauthorized - User is not authenticated" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Retrieve all users",
        "description": "Fetch all registered users. Admins can see all users; other users can only see themselves.",
        "operationId": "getAllUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "description": "Filter users by role (admin, storeowner, driver, inventoryManager, temporary)",
            "schema": {
              "type": "string",
              "enum": ["admin", "storeowner", "driver", "inventoryManager", "temporary"]
            },
            "example": "temporary"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter users by username",
            "schema": {
              "type": "string"
            },
            "example": "David Hillcrest"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": { "description": "Failed to get users" },
          "401": { "description": "Unauthorized - User is not authenticated" },
          "403": { "description": "Forbidden - User does not have access" },
          "404": { "description": "No users found in the database" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Retrieve a specific user",
        "description": "Fetch details of a specific user by ID. Admins can access all users; other users can only access their own account.",
        "operationId": "getSingleUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": { "description": "Invalid ID format" },
          "401": { "description": "Unauthorized - User is not authenticated" },
          "403": { "description": "Forbidden - User does not have access" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update user information",
        "description": "Update user information by ID. Users can update their own `preferred_name` and `phone_number`. Admins can update `roles` and review temporary accounts.",
        "operationId": "updateSingleUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "preferred_name": { "type": "string" },
                  "phone_number": { 
                    "type": "string", 
                    "pattern": "^\\+1-\\d{3}-\\d{3}-\\d{4}$",
                    "example": "+1-690-445-2710"
                  },
                  "role": { 
                    "type": "string", 
                    "enum": ["admin", "storeowner", "driver", "inventoryManager", "temporary"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "User updated successfully" },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input data" },
          "401": { "description": "Unauthorized access" },
          "403": { "description": "Forbidden - Not allowed to modify this user" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete a user account",
        "description": "Deletes a user account by ID. Users can delete their own account; admins can delete any account.",
        "operationId": "deleteSingleUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": { "description": "User successfully deleted" },
          "400": { "description": "Invalid ID format" },
          "401": { "description": "Unauthorized - User is not authenticated" },
          "403": { "description": "Forbidden - User does not have access" },
          "404": { "description": "User not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "description": "OAuth2 authentication with Google",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "openid": "Authenticate user and retrieve ID token",
              "profile": "Access user profile",
              "email": "Access user email"
            }
          }
        }
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization using Bearer token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "description": "MongoDB ObjectId" },
          "google_id": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "preferred_name": { "type": "string" },
          "phone_number": { 
            "type": "string", 
            "pattern": "^\\+1-\\d{3}-\\d{3}-\\d{4}$",
            "example": "+1-690-445-2710"
          },
          "role": { 
            "type": "string", 
            "enum": ["admin", "storeowner", "driver", "inventoryManager", "temporary"]
          },
          "date_created": { 
            "type": "string", 
            "format": "date-time", 
            "example": "2023-08-01T12:00:00Z"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}